requirements.txt
torch>=2.0
torchvision>=0.15
ultralytics>=8.0
Pillow>=9.0
numpy>=1.24
pandas>=1.5
scipy>=1.10
scikit-learn>=1.2
optuna>=3.0
tqdm>=4.64
matplotlib>=3.7
